pyproject.toml
[project]
name = "atlasai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "pydantic-ai>=1.0.10",
]
